Check grid square and foil hole registration when flushing spa
